feat: initial extraction of edit_prediction.rs
